refactor: move datasource out from resolution
