+++ internal/resolution/client/maven_registry_client.go
@@ -11,7 +11,7 @@ import (
 	"deps.dev/util/maven"
 	"deps.dev/util/resolve"
 	"deps.dev/util/resolve/version"
-	"github.com/google/osv-scanner/internal/resolution/datasource"
+	"github.com/google/osv-scanner/internal/datasource"
 	mavenutil "github.com/google/osv-scanner/internal/utility/maven"
 )
