--- internal/resolution/datasource/fixtures/maven_settings/settings.xml
@@ -1,21 +1,0 @@
-<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 https://maven.apache.org/xsd/settings-1.0.0.xsd">
-
-  <servers>
-    <server>
-      <id>server1</id>
-      <username>user</username>
-      <password>pass</password>
-    </server>
-    <server>
-      <id>server2</id>
-      <username>${env.MAVEN_SETTINGS_TEST_USR}</username>
-      <password>~~${env.MAVEN_SETTINGS_TEST_PWD}~~</password>
-    </server>
-    <server>
-      <id>${env.MAVEN_SETTINGS_TEST_SID}</id>
-      <username>${env.maven_settings_test_usr}-${env.MAVEN_SETTINGS_TEST_NIL}</username>
-      <password>${env.MAVEN_SETTINGS_TEST_BAD}</password>
-    </server>
-  </servers>
-</settings>
