--- internal/resolution/datasource/testmain_test.go
@@ -1,16 +1,0 @@
-package datasource_test
-
-import (
-	"os"
-	"testing"
-
-	"github.com/google/osv-scanner/internal/testutility"
-)
-
-func TestMain(m *testing.M) {
-	code := m.Run()
-
-	testutility.CleanSnapshots(m)
-
-	os.Exit(code)
-}
