+++ internal/scalibrextract/language/java/pomxmlnet/extractor.go
@@ -20,7 +20,7 @@ import (
 	"github.com/google/osv-scalibr/plugin"
 	"github.com/google/osv-scalibr/purl"
 	"github.com/google/osv-scanner/internal/resolution/client"
-	"github.com/google/osv-scanner/internal/resolution/datasource"
+	"github.com/google/osv-scanner/internal/datasource"
 )
 
 // Extractor extracts Maven packages with transitive dependency resolution.
